llm_pipeline/requirements.txt
requests
fastapi
uvicorn
httpx
haiku.rag
outlines 
ollama
Jinja2